Completed refactorings
