+++ pkg/injector/annotations/annotations.go
@@ -1,3 +1,16 @@
+/*
+Copyright 2022 The Dapr Authors
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
 // Package annotations contains the list of annotations for Dapr deployments.
 package annotations
 
@@ -42,7 +55,7 @@ const (
 	KeyUnixDomainSocketPath         = "dapr.io/unix-domain-socket-path"
 	KeyVolumeMountsReadOnly         = "dapr.io/volume-mounts"
 	KeyVolumeMountsReadWrite        = "dapr.io/volume-mounts-rw"
-	KeyDisableBuiltinK8sSecretStore = "dapr.io/disable-builtin-k8s-secret-store"
+	KeyDisableBuiltinK8sSecretStore = "dapr.io/disable-builtin-k8s-secret-store" //nolint:gosec
 	KeyEnableAppHealthCheck         = "dapr.io/enable-app-health-check"
 	KeyAppHealthCheckPath           = "dapr.io/app-health-check-path"
 	KeyAppHealthProbeInterval       = "dapr.io/app-health-probe-interval"
