+++ pkg/injector/sidecar/pod-patch_test.go
@@ -1,0 +1,268 @@
+/*
+Copyright 2022 The Dapr Authors
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+    http://www.apache.org/licenses/LICENSE-2.0
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package sidecar
+
+import (
+	"strconv"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	corev1 "k8s.io/api/core/v1"
+)
+
+func TestAddDaprEnvVarsToContainers(t *testing.T) {
+	testCases := []struct {
+		testName      string
+		mockContainer corev1.Container
+		expOpsLen     int
+		expOps        []PatchOperation
+	}{
+		{
+			testName: "empty environment vars",
+			mockContainer: corev1.Container{
+				Name: "MockContainer",
+			},
+			expOpsLen: 1,
+			expOps: []PatchOperation{
+				{
+					Op:   "add",
+					Path: "/spec/containers/0/env",
+					Value: []corev1.EnvVar{
+						{
+							Name:  UserContainerDaprHTTPPortName,
+							Value: strconv.Itoa(SidecarHTTPPort),
+						},
+						{
+							Name:  UserContainerDaprGRPCPortName,
+							Value: strconv.Itoa(SidecarAPIGRPCPort),
+						},
+					},
+				},
+			},
+		},
+		{
+			testName: "existing env var",
+			mockContainer: corev1.Container{
+				Name: "Mock Container",
+				Env: []corev1.EnvVar{
+					{
+						Name:  "TEST",
+						Value: "Existing value",
+					},
+				},
+			},
+			expOpsLen: 2,
+			expOps: []PatchOperation{
+				{
+					Op:   "add",
+					Path: "/spec/containers/0/env/-",
+					Value: corev1.EnvVar{
+						Name:  UserContainerDaprHTTPPortName,
+						Value: strconv.Itoa(SidecarHTTPPort),
+					},
+				},
+				{
+					Op:   "add",
+					Path: "/spec/containers/0/env/-",
+					Value: corev1.EnvVar{
+						Name:  UserContainerDaprGRPCPortName,
+						Value: strconv.Itoa(SidecarAPIGRPCPort),
+					},
+				},
+			},
+		},
+		{
+			testName: "existing conflicting env var",
+			mockContainer: corev1.Container{
+				Name: "Mock Container",
+				Env: []corev1.EnvVar{
+					{
+						Name:  "TEST",
+						Value: "Existing value",
+					},
+					{
+						Name:  UserContainerDaprGRPCPortName,
+						Value: "550000",
+					},
+				},
+			},
+			expOpsLen: 1,
+			expOps: []PatchOperation{
+				{
+					Op:   "add",
+					Path: "/spec/containers/0/env/-",
+					Value: corev1.EnvVar{
+						Name:  UserContainerDaprHTTPPortName,
+						Value: strconv.Itoa(SidecarHTTPPort),
+					},
+				},
+			},
+		},
+		{
+			testName: "multiple existing conflicting env vars",
+			mockContainer: corev1.Container{
+				Name: "Mock Container",
+				Env: []corev1.EnvVar{
+					{
+						Name:  UserContainerDaprHTTPPortName,
+						Value: "3510",
+					},
+					{
+						Name:  UserContainerDaprGRPCPortName,
+						Value: "550000",
+					},
+				},
+			},
+			expOpsLen: 0,
+			expOps:    []PatchOperation{},
+		},
+	}
+
+	for _, tc := range testCases {
+		tc := tc
+		t.Run(tc.testName, func(t *testing.T) {
+			patchEnv := AddDaprEnvVarsToContainers([]corev1.Container{tc.mockContainer})
+			assert.Equal(t, tc.expOpsLen, len(patchEnv))
+			assert.Equal(t, tc.expOps, patchEnv)
+		})
+	}
+}
+
+func TestAddSocketVolumeToContainers(t *testing.T) {
+	testCases := []struct {
+		testName      string
+		mockContainer corev1.Container
+		socketMount   *corev1.VolumeMount
+		expOpsLen     int
+		expOps        []PatchOperation
+	}{
+		{
+			testName: "empty var, empty volume",
+			mockContainer: corev1.Container{
+				Name: "MockContainer",
+			},
+			socketMount: nil,
+			expOpsLen:   0,
+			expOps:      []PatchOperation{},
+		},
+		{
+			testName: "existing var, empty volume",
+			mockContainer: corev1.Container{
+				Name: "MockContainer",
+			},
+			socketMount: &corev1.VolumeMount{
+				Name:      UnixDomainSocketVolume,
+				MountPath: "/tmp",
+			},
+			expOpsLen: 1,
+			expOps: []PatchOperation{
+				{
+					Op:   "add",
+					Path: "/spec/containers/0/volumeMounts",
+					Value: []corev1.VolumeMount{{
+						Name:      UnixDomainSocketVolume,
+						MountPath: "/tmp",
+					}},
+				},
+			},
+		},
+		{
+			testName: "existing var, existing volume",
+			mockContainer: corev1.Container{
+				Name: "MockContainer",
+				VolumeMounts: []corev1.VolumeMount{
+					{Name: "mock1"},
+				},
+			},
+			socketMount: &corev1.VolumeMount{
+				Name:      UnixDomainSocketVolume,
+				MountPath: "/tmp",
+			},
+			expOpsLen: 1,
+			expOps: []PatchOperation{
+				{
+					Op:   "add",
+					Path: "/spec/containers/0/volumeMounts/-",
+					Value: corev1.VolumeMount{
+						Name:      UnixDomainSocketVolume,
+						MountPath: "/tmp",
+					},
+				},
+			},
+		},
+		{
+			testName: "existing var, multiple existing volumes",
+			mockContainer: corev1.Container{
+				Name: "MockContainer",
+				VolumeMounts: []corev1.VolumeMount{
+					{Name: "mock1"},
+					{Name: "mock2"},
+				},
+			},
+			socketMount: &corev1.VolumeMount{
+				Name:      UnixDomainSocketVolume,
+				MountPath: "/tmp",
+			},
+			expOpsLen: 1,
+			expOps: []PatchOperation{
+				{
+					Op:   "add",
+					Path: "/spec/containers/0/volumeMounts/-",
+					Value: corev1.VolumeMount{
+						Name:      UnixDomainSocketVolume,
+						MountPath: "/tmp",
+					},
+				},
+			},
+		},
+		{
+			testName: "existing var, conflict volume name",
+			mockContainer: corev1.Container{
+				Name: "MockContainer",
+				VolumeMounts: []corev1.VolumeMount{
+					{Name: UnixDomainSocketVolume},
+				},
+			},
+			socketMount: &corev1.VolumeMount{
+				Name:      UnixDomainSocketVolume,
+				MountPath: "/tmp",
+			},
+			expOpsLen: 0,
+			expOps:    []PatchOperation{},
+		},
+		{
+			testName: "existing var, conflict volume mount path",
+			mockContainer: corev1.Container{
+				Name: "MockContainer",
+				VolumeMounts: []corev1.VolumeMount{
+					{MountPath: "/tmp"},
+				},
+			},
+			socketMount: &corev1.VolumeMount{
+				Name:      UnixDomainSocketVolume,
+				MountPath: "/tmp",
+			},
+			expOpsLen: 0,
+			expOps:    []PatchOperation{},
+		},
+	}
+
+	for _, tc := range testCases {
+		t.Run(tc.testName, func(t *testing.T) {
+			patchEnv := AddSocketVolumeToContainers([]corev1.Container{tc.mockContainer}, tc.socketMount)
+			assert.Equal(t, tc.expOpsLen, len(patchEnv))
+			assert.Equal(t, tc.expOps, patchEnv)
+		})
+	}
+}
